+++ influxql/parser_test.go
@@ -1450,7 +1450,6 @@ func TestParser_ParseStatement(t *testing.T) {
 				RetentionPolicyCreate:      true,
 				RetentionPolicyDuration:    24 * time.Hour,
 				RetentionPolicyReplication: 1,
-				RetentionPolicyName:        "default",
 			},
 		},
 		{
@@ -1462,7 +1461,6 @@ func TestParser_ParseStatement(t *testing.T) {
 				RetentionPolicyDuration:           0,
 				RetentionPolicyReplication:        1,
 				RetentionPolicyShardGroupDuration: 30 * time.Minute,
-				RetentionPolicyName:               "default",
 			},
 		},
 		{
@@ -1473,7 +1471,6 @@ func TestParser_ParseStatement(t *testing.T) {
 				RetentionPolicyCreate:      true,
 				RetentionPolicyDuration:    24 * time.Hour,
 				RetentionPolicyReplication: 1,
-				RetentionPolicyName:        "default",
 			},
 		},
 		{
@@ -1484,7 +1481,6 @@ func TestParser_ParseStatement(t *testing.T) {
 				RetentionPolicyCreate:      true,
 				RetentionPolicyDuration:    0,
 				RetentionPolicyReplication: 2,
-				RetentionPolicyName:        "default",
 			},
 		},
 		{
@@ -1495,7 +1491,6 @@ func TestParser_ParseStatement(t *testing.T) {
 				RetentionPolicyCreate:      true,
 				RetentionPolicyDuration:    0,
 				RetentionPolicyReplication: 2,
-				RetentionPolicyName:        "default",
 			},
 		},
 		{